--root now takes a qualified name (#628)
--root takes the name of the actor that shold be used as the root actor.
It used to be unqualified, like "main" and actonc would look in the
source file specified. Now that we support building multiple files in a
acton project, an unqualified name is ambiguous. Thus, by switching to
qualified name we get around this!

We still do attempt to build a qualified name if an unqualified name is
given and do this by the "old rules" so it behaves the same way as it
used to for single file compilation, for example:

  actonc examples/count.act --root main

Here 'main' will be qualified to 'count.main' and is thus equivalent to
giving the full name directly:

  actonc examples/count.act --root count.main

For project build, the root actor name must be qualified.

There are some simple test cases to check that both qualified and
unqualified names work in a project and outside.