Use common readActFile function
Rather than define how to read and parse an .act file in multiple
places, in main for a project, separately in compileFile for when we are
compiling a single file and in chaseImportedFiles, we have lifted out
that functionality to a separate function that can be reused.

The --parse arg is now consistently honoured.