Use common compileFiles for uniformity
A further simplication and unification of the compile-single-file and
project build use cases. The compileFile function which previously only
compiled a single file now takes a list of source files to compile so we
can use it both for the single file scenario as well as in a project
build.

The --verbose output was previously, due to an oversight, only available
when compiling a single file but now also works for project builds. When
used in project build setting, it only shows project level values.