Move generation of binTasks
We want to figure out our tasks and then run compileTasks on them,
rather than finding out half of the tasks (the bin tasks) during the
compilation.

I think this puts us in a better position to also do automatic detection
of root actors (matching if they are named 'main').