Straightened up parsing and type-checking of 'async' expressions.
In a term 'async E', the subterm E is now expected to be a function, not a saturated application.
The 'async' keyword binds tighter than application but less so than indexing and attribute selection,
so now 'async x.y(z)' is interpreted as '(async x.y)(z)'.

Moreover, the typing rule for 'async E' now requires the type of E to match action(*A,**B)->C, giving
the whole expression the type proc(*A,**B)->Msg[C].