Add id field to rts_descriptor
We use a / as separator unlike the DB servers where : is used. This is
partly because the host:port is a well established convention and for
RTS we don't use a port but a local-rts-id, so we don't want people to
associate this with a port when they see it. But also because we will
include both DB servers and RTS nodes in the same lists in the
monitoring interface and by using a different separator we avoid
collisions.