Expose list of actors over RTS mon interface
It is now possible to display the actors in the system. They are exposed
over the RTS mon interface through the "actors" query.

actonmon supports displaying them both in a rich table or as more basic
output without requiring rich being installed.

Unlike the other JSON output, we here create an artificial key since the
actor_id struct doesn't have a char key and we didn't feel like adding
one. Downside is more memory usage while running the mon command but on
the other hand the memory usage is reduced in other places.