Test compilation of examples & clean up examples
We now test that all the example applications we have in the examples/
directory actually compile. Testing that the program run correctly would
require writing individual tests for each program which would be very
time consuming and probably not worthwhile.

Many examples are broken or are really more of tests. I've reorganized
some so that the ones that are clearly tests are now in test/builtins
while some others, whose purpose is not entirely clear and should
probably be cleaned up a bit, have been moved to workspace/misc-tests.

All examples now compile and must henceforth compile for this test suite
to pass.

The examples are not compiled with a root actor, so no binaries are
produced which means we don't run the linker and thus will miss certain
errors. Once we add automatic root actor detection (if there is an actor
named "main") it is likely we will see more errors as I believe some of
the examples are currently broken.