Print debug output on test failures
Tasty doesn't seem to have a built-in way of printing some output of
tests but only when the test has failed. Likely because unit tests etc
normally doesn't run a program under the hood etc, they are much smaller
and have more deterministic failures.

Anyway, I've fixed it the pragmatic way here by passing around more
information. If we see a failure and we don't expect it, then we just
print stuff to stdout. The output of the test script will be somewhat
garbled, but this is essentially impossible to avoid since we are also
running the tests threaded.

Through the process, I've restructured quite a bit of the code and it's
now much easier to build up the test cases, simply calling the testFiles
function with a name for the test group, a directory and a list of
expected failures.