Simplify test by combining file & project functions
The inner functions, buildFile and buildProject, used to specific to
compiling a file or project. This means all wrapping functions we add on
top typically need to be duplicated as well.

By determining if something is a file or project simply by looking at
it (is it a directory=project or not=file) we can unify it to a single
buildThing function and also have combined wrapper functions. Much
better.