Support test wide failure expectation
The implicit expectation has been that tests would succeed and we
provide a list exceptions for tests that currently fail. However, for
some tests we expect them all to fail (based on directory we are looking
in) so rather than having to list them all, we want to set an
expectation for all to Fail. That is now possible.