Support Acton projects as tests
We can now place standalone .act files or Acton projects in the
respective test directories, like in test/regression/ which is nice
since some tests are better organized as a project!

We used to search recursively for tests in the provided directory but we
never really used that and it breaks if we try to put a project in there.

A very basic sanity check is performed that a directory also contains an
Acton.toml file to be considered a project.