Move test of async context
We inspect the output of this test and so have to manually define a
testCase for it rather than bunch it together with the rest of the
discovered tests. Wrote a few convenience functions for buildAndRun
similar to what we have with buildThing.

I think this is part of the core language, so created a directory for
that. My vague idea is to have like core, builtins, rts, ddb and our
existing regressions or something like that.

This is the last remaining of our tests with special handling, the whole
regression testing is now moved to Tasty!