Do not link numpy with math.o at build time
The reason we link numpy with stuff in the first place is because it has
external dependencies, like libbsd, libdl and libml and if we don't link
them in numpy.o at build time, actonc would effectively depend on those
libraries at run time. We don't want run time dependencies on libraries!

I'm not sure why math.o was added too but I am guessing that I saw some
error when only providing parts of the libraries!? Can't reproduce it
now and removing it seems to work, so all good?