Add more test categories
To simplify testing, the idea is to categorize all tests into test
groups by the "component" they test. This is so that when you are
working on e.g. the RTS, you can run tests for the RTS and not have to
wait for the whole test suite to complete. There are corresponding make
targets for each category:
- builtins: make test-builtins
- compiler: make test-compiler
- core language: make test-lang
- database: make test-db
- examples: make test-examples
- rts: make test-rts
- stdlib: make test-stdlib
- regressions: make test-regressions

regressions is last because it's not a component category but rather the
kitchen sink where things go which don't fit anywhere else. Avoid
placing things there unless you don't find a better place.

Any .act file or Acton project directory placed in there will be
discovered by the test system and included in the corresponding test
group. By default tests programs are expected to compile and run
successfully (exit 0). It is possible to use a file suffix to denote a
different expectation, namely:
- __bf means build failure and sets an expectation that actonc will exit 1
  - for example `foo__bf.act` is expected to fail actonc compilation
- __rf means run failure and sets an expectation that actonc will exit 1
  - `bar__rf.act` is expected to compile successfully but to return with
    exit 1 when the program is run

Moved some tests around to fit into the new categories and improved
comments in the abstract class tests.