Do not link numpy with math.o & no common symbols
The reason we link numpy with stuff at its compile time is because it
has external dependencies like libbsd, libdl and libml and if we don't
link them in at build time of numpy.o, actonc would effectively depend
on those libraries at run time! We don't want run time dependencies on
libraries!

I added math.o to work around some issue in the past but that in turn
caused problems of duplicate symbols if a program imported both math
and numpy.

It is now fixed in what I believe is the proper way. math.o is not
linked into numpy.o. They are linked together by actonc if necessary.
The missing piece on MacOS / clang is to use -fno-common, which gcc
seems to do per default.