Add libuv to Acton system
This adds libuv to the Acton system.

We need to keep a copy of the libuv static lib archive in the Acton
system distribution, so we copy it from the OS system distribution at
build time of Acton.

This way we don't have to deal with the complexicty of building libuv
ourselves but can rely on our OS / distribution. Our build times will
naturally be lower as installing a pre-built binary package is faster
than compiling it from scratch. On the downside, we don't get to version
pin libuv nor we do have any say in how it is compiled. We could imagine
that we would want a _dev and _rel version of it.

If we think the downsides increase in importance or we encounter issues
due to the lack of pinning, we can always choose to include libuv
sources in acton and build ourselves at a later point in time. This
keeps us smaller and simpler for the time being.

The include statements of uv.h in RTS and elsewhere are placed so that
uv.h is only needed at compilation time of the Acton system, not at
actonc run time. This way we do not need to include the libuv headers in
the Acton system distribution. This does presume that anythin in Acton
using libuv will be compiled as part of the Acton system distribution,
which is probably true today, but in a longer term future it could be
that we want to allow users to write their own modules in C and hook
into libuv. At that point it might be better if we ship the headers too?
We'll cross that bridge when we get there.