Select single linux tar ball for release
Now that we run the test suite on multiple Linux distros we produce an
artifact for each which confused the release jobs. Pick a single one,
the one built on debian-11, for the release artifact. The rest are still
available but not as release assets.

Restructured the matrix with separate os and version variables so that I
can construct a compound value without a colon in it since that is
forbidden in upload-artifacts.