Ensure dependencies are included & specified
Add new test job running on debian:11. Pull down .deb file produced in
earlier job, install and run actonc on a simple program to ensure
dependencies are correctly specified.

We have dependencies at compile time of actonc itself and then a
different set of dependencies at run time of actonc (when a user is
compiling a program).  We want to minimize the number of dependencies we
have at run time. In order to ensure we get this right, we have to
install on a fresh machine that does not have the actonc compile time
dependencies installed. Further, for the dependencies we do have, we
try to include them in the Acton system distribution itself.