Add failing test case for rebuilding a project
There's a problem with the new root actor discovery mechanism. When a
project is rebuilt, actonc crashes with:

    kll@ThinkYoga:~/acton/test/compiler/rebuild$ ../../../dist/bin/actonc build
    Compiling rebuild.act for release
    Building executable /home/kll/acton/test/compiler/rebuild/out/rel/bin/rebuild
    actonc: Maybe.fromJust: Nothing
    CallStack (from HasCallStack):
    error, called at libraries/base/Data/Maybe.hs:148:21 in base:Data.Maybe
    fromJust, called at ActonCompiler.hs:570:43 in main:Main

Which I'm guessing has to do with that we do not read in all modules
from .ty files unless we need to. When building from scratch, all
modules are read in, since we are compiling them all, but when some are
already built, we skip them unless they are dependencies of other
modules.