Add net module with basic DNS lookup functions
This adds a new module called 'net'. The intention is to collect all of
the network related things currently existing in the env actor and place
them in the net module instead and perhaps to expand it further.

The only functions currently added are for doing DNS lookups of A and
AAAA records. The interface is not really what we want in the end but
quite similar to what we currently have.

The module is built using the new external C function style, which means
that functions in the .act file whose body consists of a NotImplemented
statement are not generated by the compiler but rather expected to be
manually or EXTernally implemented in the file net.ext.c, if such a file
exists.

libuv is used for these new functions also marking our first use of libuv.

lookup_a and lookup_aaaa (the public functions in the net module) take
an auth argument, which is expected to be a DNSAuth token, which in turn
must be created from a NetAuth token. The idea is to have a WorldAuth
token fed it to the root actor but we're not there yet, so the NetAuth
token currently accepts a None token as input. That should change once
we have the other bits in place.