Detect root actor in tests
We used to pass in --root main to actonc to get it to build a binary but
now that we have automatic detection of the root actor, that is no
longer necessary. In particular for projects, which require a qualified
root, this simplifies things as we can now just build a project and get
binaries built for all files with a actor named 'main'.