Support running tests in projects
Thus far the only test projects we've had have failed so we never got to
running the test program and expecting a success.

runThing now supports running an executable in a project. It assumes the
binary executable is named the same as the project without our test
expectation ending. For example, if a project is named import_actor__bf
then we will try to run ./import_actor in import_actor__bf/out/rel/bin