Join time & _time module
We used to have these split because at the time we figured that we would
keep C code in the _time module and wrap it up in Acton code in time.
Since then we have come up with the new concept of using NotImplemented
and .ext.c so we can now rather easily combine C and Acton code in the
same module!