Force recompilation for --sigs & --[ch]gen
There are various arguments to the compiler, like --kinds or --norm that
produces debug output from the different compilation passes inside of
the compiler. These are side effects from the compilation which means
they are subject to the up-to-date check, that is, if a file is deemed
up to date, then it will not be compiled and if it is not compiled,
there won't be any side effects run either. From a user perspective,
this is quite surprising as --deact will produce no output if it's
already been compiled. Now applying any of these flags will force
recompilation of the module and thus to produce the output.

It could be argued that the dump commands should be used to view this
information, but it's not the same, since it can only show the content
of a .ty file and simply doesn't have information from other compiler
passes. Also, it would require running two commands, first to build a
module and then dump to see output.

In addition, for --cgen and --hgen, the primary output and thus the
primary job of the compiler is not really to compile a module but rather
to produce the requested output. The intent has changed and thus, for
both --cgen and --hgen the compiler now exits after printing the code,
and does not continue with compilation.