Compile with relative paths for pretty log paths
Our logging functions include the path to the source code. It uses the
__FILE__ macro to achieve this and it is defined by which filename is
fed into cc. If we invoke cc with absolute filenames, we get absolute
filenames in __FILE__ and thus in our log messages, like so:

    10:48:53.701706 DEBUG /home/kll/terastream/acton/stdlib/src/process.ext.c:   79: RTS  3: Pipe init stdin handle 0x7f2860001318

Somewhat shorter relative paths are much more elegant. We still have a
small challenge in that we compile both the system as well as some acton
project so we want to be able to differentiate between these. Thus, the
solution is to use relative paths that start at the parent directory of
the current project. Our stdlib is also organized as a project, so it
will now look like:

    15:25:32.623690 DEBUG stdlib/src/net.ext.c:   42: RTS  4: HELLO A

And similarly for a module in a users project will have that project
name in the path.

We could go even further, trying to come up with some special form of
compilation for stdlib and something special for user projects to make
even better looking paths, but this felt like a good balance that's easy
enough to achieve.