Await ioloop startup in RTS
We have to wait for the ioloop to start before doing other things. If we
try to register IO callbacks before the ioloop has started properly
we'll get undefined behavior.

This is a rather crude hack but it seems to work rather well. We should
think about how we can restructure the RTS so this becomes more elegant.
Perhaps the main thread should be the libuv thread?