Proper time module
This significantly extends the time module with a proper, somewhat
thought through design of how to work with time in Acton. It boils down
to two usage patterns:
- measuring elapsed time (e.g. for benchmarking)
- telling the current date and time (e.g. for logging)

We want to design a user experience that makes this straight forward and
that deliver correct results. Time is hard and most people have no idea
how complex it is.

For measuring elapsed time, we offer the Stopwatch. Usage is simple:

    s = time.Stopwatch()
    # do something
    elapsed = s.elapsed()
    print("Elapsed time:", elapsed.str_us())
    # Elapsed time: 1.234567 s

Most languages expose some form of timestamp object so the user can grab
a t1 timestamp before and t2 after the thing they want to time. Then
it's a simple matter of subtracting t1 from t2 and you get the diff. A
big problem with this is that you could also imagine using a "timestamp"
for telling the time, like using for a log. This sort of interface where
the user is working with raw timestamp fails to capture the *intent* -
does the user want to tell the time or measure elapsed time? Not so with
a Stopwatch - it is utterly clear that it is about measuring elapsed time.

Further, since Acton supports actor migration, which is fairly unique in
the computing world, this poses unique challenges. Any normal program is
bound to terminate when the computer it is running on shuts down, which
also means that the monotonic clock is always available to the program.
An actor in an Acton program however can outlive a single computer by
migrating to run on a different computer, thus seeing different
monotonic clocks. To combat this, the Stopwatch can fall back to using a
measurement from the realtime clock instead. It is then bound to the
precision of that clock.

In order to convey precision, all time measurements, like Instant and
DateTime carry clock information so that we can produce a result with a
known precision and error.

For telling the time, the time.now() or time.utcnow() function is the
primary interface, which returns DateTime object. It can easily be
formatted as a ISO8601 string.

Many things are missing but I think this is a good start and above all
it feels like a solid foundation in terms of user experience.

There are various bugs in the compiler or language shortcomings (no u64
for example) so there are a few workarounds in the code that we need to
sort out but the overall module design is pretty set I think.