Use new ext style for random & acton.rts modules
Rather than have the whole module implement in C with a .c file, we now
use the newer ext.c style to implement certain functions (which in this
case happens to be 100% of the module contents but that's besides the
point).

ext.c do not require manual Makefile targets, so those have been removed.

Also removed unused _time make targets.