Inspect all source files in up-to-date check
We now inspect the modification times of all source files and not just
the .act file to determine if a module is up to date or not. Previously,
we would miss compiling a module if it was implemented in C and only the
mod.(ext.)c file was modified.

There is no deeper analysis of C source code, so if mod.c imports foo.c
which in turn has been modified, that is not visible to the compiler and
thus won't lead to compilation. If a custom Makefile is used, we also
have no visibility into dependencies, so the source file will need to be
touched to force compilation on every change.

Further, paths have been fixed when checking output files. I
accidentally broke this when I changed to using relative paths in most
places. Whenever we do IO calls, like doesFileExist, we should prefer
absolute paths!