Always compile custom make targets
Most of the compilation of acton code happens in runRestPasses which is
conditioned on an up-to-date check, so that if the output is up to date,
we do not run any compilation. However, with custom make targets, which
we use for modules implemented entirely in C, we don't track all output
files, since the acton compilation is limited to stub mode. Thus, to
ensure object files are built we now always run the custom make targets.