Upgrade stack on old dists
The haskell stack in Ubuntu 20.04 is too old to properly function. We
need to upgrade it first!

Also install libmd since our numpy module depends on it. We haven't had
to explicitly list it since on newer versions of ubuntu & debian, it is
pulled in as a dependency of some other package we install. However, on
Ubuntu 20.04 that doesn't happen, so it needs to be explicity listed.