Combine IO & worker threads
What used to be the worker threads are now instead a combined type of
worker thread that also does IO!

libuv now sits at the center of each thread and running actor
continuations is a function that is invoked from the libuv loop when
needed.

All IO actors that call libuv functions MUST first pin their affinity to
a particular worker thread!

There are multiple reasons for this design. We have been pondering and
discussion, for quite some time, the merits of a "shared nothing"
architecture where worker threads are able to do work without sharing
any information (and thus avoiding locks) to the highest possible
degree. This is certainly a move in that direction! However, what
ultimately forced this work right now is simply that for libuv to work,
all calls to uv functions must happen from the libuv thread. An actor
continuation being run by a worker thread thus cannot make uv calls for
a single shared IO thread. I believe the idiomatic UV design pattern is
to put things on a queue and wake up the IO thread so it can pick up the
work from the queue. We already have a queue, or multiple thereof; the
readyQs! Thus, with this new design we can schedule actors that are
doing IO on a particular thread and can then run libuv functions
directly!

A potential downside of mixing IO with running actor continuations is
that a long running continuation can starve out IO. There is no limit to
the run time length of a continuation, so this could become a real
problem. Doing preemption could be a path forward but the simpler
solution would be to divide up so that only some threads do IO and then
only process IO actors while the rest of the threads would pick up any
actor continuations.