Combine IO & worker threads
What used to be the worker threads are now instead a combined type of
worker thread that also does IO!

libuv now sits at the center of each thread and running actor
continuations is a function that is invoked from the libuv loop on every
iteration. It tries to run for a maximum of 20ms before yielding to IO,
however since we are not doing preemption, a single long continuation
can block for an arbitrary amount of time. An async callback is used to
"wake" up a loop that is currently blocking waiting on IO events.

All IO actors that call libuv functions MUST first pin their affinity to
a particular worker thread!

There are multiple reasons for this design. We have been pondering and
discussion, for quite some time, the merits of a "shared nothing"
architecture where worker threads are able to do work without sharing
any information (and thus avoiding locks) to the highest possible
degree. This is certainly a move in that direction! However, what
ultimately forced this work right now is simply that for libuv to work,
all calls to uv functions must happen from the libuv thread. An actor
continuation being run by a worker thread thus cannot make uv calls for
a single shared IO thread. I believe the idiomatic UV design pattern is
to put things on a queue and wake up the IO thread so it can pick up the
work from the queue. We already have a queue, or multiple thereof; the
readyQs! Thus, with this new design we can schedule actors that are
doing IO on a particular thread and can then run libuv functions
directly!

A potential downside of mixing IO with running actor continuations is
that a long running continuation can starve out IO. There is no limit to
the run time length of a continuation, so this could become a real
problem. Doing preemption could be a path forward but the simpler
solution would be to divide up so that only some threads do IO and then
only process IO actors while the rest of the threads would pick up any
actor continuations. Since all IO actors are written by Acton
maintainers, we could achieve a soft guarantee on scheduling delay
simply by ensuring that no continuations are very long. It will
certainly be good enough for this.

The prep callback was specifically selected so that we would always
check for work after IO callbacks has been called, since it is very
likely that there is work for some actor after an IO callback has run.
An earlier attempt only made use of an async callback for doing the
work, but it placed the responsibility of waking up one or more worker
threads on the author of the IO module, for which there isn't a good
interface - per design, as it's supposed to be abstracted away. Checking
if there is some work should be very cheap anyway since we have a
shortcut that doesn't even take a lock!