Add Process.signal(), .stop() & friends
This makes it possible to send signals to a running process, including
well known signals like term() and kill() or any custom ones through
signal().

It adds a standard way of stopping a program through the .stop() action
which first attempts to stop a process by sending SIGTERM followed by a
SIGKILL after 1 second.

This is currently not portable, we rely on the signal numbers used on
UNIX systems. In some future, we should adapt this to work across other
OSes.