Use on_exit callback for errors
Rather than raising an exception when we encounter an error in spawning
the process, we call the on_exit callback. It now receives a new
argument which is the error message. If there is no error and the
process exited for some other reason, the error variable is None.

The reasoning for this approach is that it's likely we need to write
less code and have a more consistent control flow when using callbacks.
If a process exits abnormally, we would write a on_exit callback to
handle this and perhaps restart it. If a process doesn't start / spawn
properly, that is sort of in the same category of events and should
probably lead to a restart.