Remove arg from pin_actor_affinity
This makes it easier to call, since we can do it from any function and
not just an actor. All functions are called by an actor anyway, it's
just that we dig it out of the pthread store rather than having to
thread through the argument (which is only possible in C anyway, making
it impossible to call such a function directly in Acton).