Add TCP client actor net.TCPIPConnection
The net.TCPIPConnection actor is a TPC client which connects to remote
systems. It takes an IP address as argument. The idea is to wrap this up
in a net.TCPConnection which does name resolution including happy
eyeballs.

It handles resume for distributed DB systems, which is also verified by
a test. A few tests have been updated to use this rather than the old
TCP client connection actor.