Ensure we grab latest artifact
There could be multiple artifacts when there have been failures during
the job that produced it. The artifacts are uploaded immediately after
the build completes so it can upload and then fail in the subsequent
testing, the job is rerun and thus we get multiple artifacts.

This makes the pre-release-tip handle that scenario by picking the
latest available artifact.