Add JSON module
This adds a JSON module with encoding and decoding abilities. There's a
json.Json actor that does the actual work. I was originally going for
just having a free decode and encode function in the module but that
doesn't work (compiler bug) so had to make it look like an actor, which
does bring the benefit of allowing async work. There are free
encode/decode functions, so that you can encode or decode JSON with a
single line of code. They just wrap up the Json actor so every
invocation also means creating an instance of the Json actor. It should
really be the other way around, but we'll have to sort out the compiler
bug first.