Make custom make more idempotent
By always copying the .h file from src/%.h to out/types/%.h, any custom
make targets would always think the source was fresher than the output
.o file and thus run. We now avoid this by first comparing the source .h
file with the target one.

On a AMD 9550 it takes 750ms to build numpy.o, and in total for all
stdlib modules somewhere north of a second, which doesn't sound like
much but is 10x more than with this fix. On slower machines it is even
more noticeable.

I don't think there are any adverse performance consequences from this.
Our header files shouldn't typically be that large, so this shouldn't be
much of an issue. Anyway, it is still cheaper to compare two files than
recompiling a large module.