after now takes float seconds as argument
The after statement has thus far taken an integer argument for the
number of seconds after which to run a method. It is now changed to a
float so we can do sub-second delays.

Remove old test case that showed how we do not automatically convert
integers to float. Since it's not really a bug but rather a design
limitation we don't need a regression test for it.