Retry DB tests up to 5 times
We have a lot of intermittent failures in the DB tests, most of which
stem from segfaults. Manually rerunning failed CI jobs is boring and
time consuming. Also, retrying means rerunning the entire job which
takes 5 minutes while retrying a single DB test is on the order of
seconds, so this is much better.

In general we've stayed clear from retrying tests as to avoid hiding
intermittent failures but it's clear here that we'd be more productive
if we just solve the test-retry issue in a pragmatic way. We still have
tickets and won't forget to get rid of the errors in the DB stuff.