Make DB test retry actually work
While we did retry the test function we have basic state in our
environment that is automatically set up in the setUp function and later
torn down in tearDown(). The retry wrapper wouldn't rerun those methods.

Instead, the test is refactored slightly and we explicitly call
dbc_restart() from the test function to ensure the DB cluster is up and
running in a fresh state.