Include and use runacton in tests
Forgot to actually include runacton into our distribution. That's fixed
now and in order to properly test that, we now use it in both the
run-linux and run-macos tests!

On mac we need to modify the PATH since we don't actually install acton
on the system but just extract it in a directory in our home dir. It
would be preferable if we actually did a brew install or somesuch so
that acton would be installed in the system...