Uniformly use 'macos' in CI test job names
We had a mix of macos & darwin. It's now more uniform.

'darwin' comes from the architecture, i.e. uname reports itself as
darwin so that is what our package names are called but still, it's
easier to understand what it is if we refer to it as macos in most
places I think...