Add Homebrew test build
The authoritative source for our Homebrew Formula is now in this
repository rather than in homebrew-acton. We've had this long standing
issue of the homebrew-acton repository becoming outdated and there is no
good feedback path into this repo. This is an attempt to address that!

The Formula itself is updated to reflect our current dependencies.

There's a new CI job that runs something similar to what the
homebrew-acton CI pipeline looks like. We can't do the exact same thing
because of how limited the brew test-bot is. Rather stupid if you ask
me... but anyway, we trick it a bit to make it run a test build of acton
from our current branch, so we'll see if the Formula is correct on a PR
for example.

A second CI job tests installing the produced bottle.

The brew job is only run on macos-12. For some weird reason, macos-12 is
quite consistently faster than macos-11. The number of concurrent builds
on macos is limited by GitHub, so we save on those resources a bit by
only running this job on macos-12. If even that proves to be too much,
we can limit this job to only run for releases or PRs for releases.