Add auto-release CI job
Releasing thus far has consisted of:
- updating the VERSION variable common.mk
- updating CHANGELOG.md
- pushing a tag

It's slightly annoying because the first two means changing files,
opening PR and merging and then git pulling main, add a tag and push the
tag. It's a lot back and forth.

We now automate the last bit of pushing a tag by having a new CI job
doing it for us.

New process is thus:
- updating the VERSION variable common.mk
- updating CHANGELOG.md
- open PR with name starting with release-
  - when PR is merged, tag based on VERSION in common.mk will
    automatically be pushed to main branch