Use deploy key for pushing tag
The push used the defaults credentials provided when checking out the
repository, which equates to github.token. For some reason, perhaps to
avoid cyclic errors, the CI runner is not supposed to trigger itself and
so using github.token for auth will not trigger any workflows. By
switching to our own SSH deploy key for checking and pushing the repo we
can trigger workflows!