Default gossip port to RPC port +1
The gossip protocol only includes the RPC port and does not explicitly
say what port is used for gossip, so other parts of our system assumes
it is the RPC port +1. Thus, we change the actondb default to also be
based on RPC port +1.

The long term solution is to make the gossip port configurable and to
propagate that information in the gossip protocol. Meanwhile, since we
do not support any other port than +1, we emit a log error on this
invalid configuration but leave most of the argument parsing code in so
that it's easy to change one day.