Properly initialize WT state
When the worker threads are created they are first in a no-exist / poof
state and once they're initialized in the main_loop function they should
be marked as Idle. They were not, which lead to problems as the wake_wt
function specifically looks for worker threads in the idle state to wake
up. Since it was in poof state, it would not be woken up. Now fixed!