Only communicate with live nodes
The DB client interaction functions have unconditionally tried to
communicate with all remote servers in our list. If a server is dead,
it's socket is invalid and trying to communicate with it doesn't make
any sense. We now only communicate with servers that we think are live.

This has been a long running issue and a rather annoying one where we've
seen lots of garbage written to the screen when RTS fails to connect to
a DB server. It happened because the sockfd gets rewritten to 0, which
is stdin, and then the message meant for a DB server is written to our
terminal. It's a binary protocol so it looks like a bunch of garble
garble.

Now also using -1 for sockfd so we get an error if we try to write to it
and tightened up so we only set NODE_LIVE after connection is successful.

Same changes also goes for the actondb server which has had similar
problems with garbage output in the past.