Invalidate fds on resume
After we resume a TCP actor, its fd is invalid. While we do already call
the registered error callback to notify the user application and let
them deal with it (reconnecting), there are probably cases where a
method is already scheduled to run and it will attempt to use the
invalid fd. By setting it to an invalid fd, -1, we should see errors
reported (provided that all our calls to uv_write and similar are
correctly error handled) instead of segfaults.