Rewrite TCP server resume from DB test
In this test where we start up a DB cluster and then run a RTS node
against it, running a TCP server program, we want to be able to react to
what is happening with all involved components. Callback style IO, like
in Acton, is particularly suited to this and so I've rewritten the test
in Acton. While I'm very happy with how this test program performs, like
how it is able to immediately react to all sorts of input events, I'm
not quite happy with the look of the test program. It's not very easy to
follow at all but should do the job for now. We'll find the time to come
up with a better design pattern. It's much faster to run this than the
previous test_db.py script.

One of the problems with the test_db.py script has been the lack of
output during failures. This has been addressed so we collect the output
of all programs (DB servers & main app under test) and output this to
screen, inluding a short prefix, if the test fails. I find this very
simple to follow.