Only check for fd == 0
We might actually end up with invalid fds, like -1, since we don't
mutex protect everything.. and that should be fine, so changing the
assert accordingly.

Initially, 0 was used as an error signal. That was wrong since 0 is
actually a valid fd and when we meant to signal "invalid fd" we just
ended up writing data to stdin which in turned led to garbage output on
terminals... anyhow, we changed the "invalid fd" to use -1, so this
assert is really just to check that we haven't forgotten anywhere.